requirements.txt
Django==5.2.3
pillow==11.2.1
environs==14.2.0
marshmallow==4.0.0
requests==2.32.4
django-admin-sortable2==2.1.3
django-tinymce==4.1.0